Simplify the CPT
Since the CPT isn't really trested like a normal post, we can simlify
it a bit since it really only needs to be saved. Helps fix confusion
about creating an album and having to have it be published in order to
have it be in a post.

All albums can be published at their creation, this does not put them
anywhere on the frontend because they have to be added via a shortcode.