Flip rows properly.
This fixes an issue with flipping rows on the Album images metabox when
editing a photo album. Previously we were flipping the order, but that
is not necessary since the rows are already in the right order.

We simply needed to switch the rows content (HTML) and switch the ID's.